+++ src/Common/UpdaterCore/Updater/UpdaterInstaller.cs
@@ -25,5 +25,6 @@ public class UpdaterInstaller : IWindsorInstaller
     public void Install(IWindsorContainer container, IConfigurationStore store)
     {
         container.Register(Component.For<UpdaterResourceSetBuilder>().LifestyleSingleton());
+        container.Register(Component.For<UpdaterHash>().LifestyleSingleton());
     }
 }